+++ understory_virtual_list/CHANGELOG.md
@@ -8,10 +8,14 @@ Subheadings to categorize changes are `added, changed, deprecated, removed, fixe
 
 # Changelog
 
-The latest published Understory Virtual List release is [0.1.1](#011-2026-05-17) which was released on 2026-05-17.
-You can find its changes [documented below](#011-2026-05-17).
+The latest published Understory Virtual List release is [0.1.2](#012-2026-05-22) which was released on 2026-05-22.
+You can find its changes [documented below](#012-2026-05-22).
 
 ## [Unreleased]
+
+## [0.1.2][] (2026-05-22)
+
+This release has an [MSRV][] of 1.88.
 
 ### Added
 
@@ -72,7 +76,8 @@ This is the initial release.
 [#170]: https://github.com/forest-rs/understory/pull/170
 [#171]: https://github.com/forest-rs/understory/pull/171
 
-[Unreleased]: https://github.com/forest-rs/understory/compare/understory_virtual_list-v0.1.1...HEAD
+[Unreleased]: https://github.com/forest-rs/understory/compare/understory_virtual_list-v0.1.2...HEAD
+[0.1.2]: https://github.com/forest-rs/understory/compare/understory_virtual_list-v0.1.1...understory_virtual_list-v0.1.2
 [0.1.1]: https://github.com/forest-rs/understory/compare/understory_virtual_list-v0.1.0...understory_virtual_list-v0.1.1
 [0.1.0]: https://github.com/forest-rs/understory/releases/tag/understory_virtual_list-v0.1.0